+++ src/main/java/com/mallang/bobby/domain/auth/oauth2/dto/AccessTokenResponse.java
@@ -1,4 +1,4 @@
-package com.mallang.bobby.oauth2.dto.kakao;
+package com.mallang.bobby.domain.auth.oauth2.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -8,7 +8,11 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-public class KakaoUserResponse {
-    private Long id;
-	private KakaoAccount kakaoAccount;
+public class AccessTokenResponse {
+    private String tokenType;
+	private String accessToken;
+	private Integer expiresIn;
+	private String refreshToken;
+	private Integer refreshTokenExpiresIn;
+	private String scope;
 }
