+++ src/main/java/com/mallang/bobby/domain/auth/oauth2/dto/OAuth2Provider.java
@@ -1,4 +1,4 @@
-package com.mallang.bobby.oauth2.dto;
+package com.mallang.bobby.domain.auth.oauth2.dto;
 
 import lombok.Getter;
